terminate instance on unexpected exception

+++ src/main/java/com/veertu/plugin/anka/AbstractAnkaSlave.java
@@ -68,8 +68,8 @@ public abstract class AbstractAnkaSlave extends AbstractCloudSlave {
 
     public void terminate() throws IOException, InterruptedException {
         super.terminate();
-        vm.terminate();
-
+        if (vm != null)
+            vm.terminate();
     }
 
     @Override
